requirements.txt
streamlit==1.32.2
Pillow==9.4.0 # For PIL, the Python Imaging Library
numpy==1.24.3
tensorflow==2.16.1
streamlit_image_select==0.6.0
pandas==2.0.3
